refactor: extract diagnostics routes from resumes.ts (4112→3969 lines)
Move 4 diagnostic route handlers (analysis-tasks, skills-version,
field-coverage, list-diagnostics) and their helper function into a
new resumes_diagnostics.ts module. Mounted before resumesRoutes so
static diagnostic paths are not shadowed by the {resumeId} dynamic
route in resumes.ts.

Original: 4112 lines in resumes.ts
After:    3969 lines in resumes.ts + 171 lines in resumes_diagnostics.ts
